+++ core/custom_transport.go
@@ -1,7 +1,9 @@
 package core
 
 import (
+	"context"
 	"crypto/tls"
+	"net"
 	"net/http"
 	"time"
 
@@ -9,6 +11,25 @@ import (
 )
 
 const defaultRetryInterval = 1 * time.Second
+
+var socketTransport = &http.Transport{}
+
+func init() {
+	socketTransport.DisableCompression = true
+	socketTransport.DialContext = func(_ context.Context, _, _ string) (net.Conn, error) {
+		return net.DialTimeout("unix", "/var/run/docker.sock", 32000000000)
+	}
+}
+
+type SocketPassthroughTransport struct{}
+
+func (r *SocketPassthroughTransport) RoundTrip(request *http.Request) (*http.Response, error) {
+	if request.URL.Host == "/var/run/docker.sock" {
+		return socketTransport.RoundTrip(request)
+	}
+
+	return http.DefaultTransport.RoundTrip(request)
+}
 
 type retryableHTTPTransport struct {
 	transport http.RoundTripper
@@ -28,7 +49,11 @@ func (r *retryableHTTPTransport) RoundTrip(request *http.Request) (*http.Respons
 func (r *retryableHTTPTransport) SetInsecureTransport() {
 	transportClient, ok := http.DefaultTransport.(*http.Transport)
 	if !ok {
-		logger.Warningf("cli: cannot use insecure mode with DefaultTransport of type %T", http.DefaultTransport)
+		logger.Warningf(
+			"cli: cannot use insecure mode with DefaultTransport of type %T",
+			http.DefaultTransport,
+		)
+
 		return
 	}
 	if transportClient.TLSClientConfig == nil {
